Api/requirements.txt
anyio==3.7.0
click==8.1.3
defusedxml==0.7.1
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tabula-py==2.7.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
